Krish Naik - OneShot/requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
google-generativeai

langserve
fastapi
uvicorn
sse_starlette
